requirements.txt
click==7.1.2
Flask==1.1.2
greenlet==1.0.0
importlib-metadata==4.0.1
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
SQLAlchemy==1.4.11
typing-extensions==3.7.4.3
Werkzeug==1.0.1
zipp==3.4.1
